move pprof internal api to separate file
